Added a DeclareQueueWithCounts() function
Currently, SimpleAmqpClient has no means to check consumer and message
counts in a declared queue, which is required in some cases (e.g. to
display these values or prevent a queue to get too large). Added a new
function which returns these figures, and changed the DeclareQueue()
functions to call the new one so there is no duplicate .