lib: refactor Channel ctor and Create funcs
This does a better job of hiding implementation details and better
prevents users from accidentially taking a dependency on the internal
implementation details about how the Channel constructor works.

This may be a breaking change for those that have used one of the
existing constructors directly instead of one of the Create factory
functions.

This sets the groundwork for an upcoming API improvement of this
interface.