Fix setting of flags in CreateAmqpProperties().
We cannot just replace the entire `ret._flags` variable when setting certain
message properties. We need to OR the variable with the corresponding flag.
Otherwise, setting e.g. message expiration resets all the previously set flags,
such as for priority or delivery mode. Many flags are ORed correctly (e.g.
priority or delivery mode), but some are not. This commit fixes that.

The bug that this commit fixes was introduced in commit
938b102f8fc85e5113b3e55700b8af9b71d9f01d.